split out base object traits
