+++ gobject-base/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "gobject-base"
+version = "0.1.0"
+authors = ["Sebastian Dröge <sebastian@centricular.com>"]
+repository = "https://github.com/sdroege/gst-plugin-rs/tree/master/gobject-base"
+license = "MIT/Apache-2.0"
+
+[dependencies]
+libc = "0.2"
+lazy_static = "1.0"
+byteorder = "1.0"
+glib-sys = { git = "https://github.com/gtk-rs/sys" }
+gobject-sys = { git = "https://github.com/gtk-rs/sys" }
+glib = { git = "https://github.com/gtk-rs/glib" }
+
+[lib]
+name = "gobject_base"
+path = "src/lib.rs"
