+++ gst-plugin/Cargo.toml
@@ -15,6 +15,7 @@ keywords = ["gstreamer", "multimedia", "audio", "video", "gnome"]
 libc = "0.2"
 lazy_static = "1.0"
 byteorder = "1.0"
+gobject-base = { path="../gobject-base" }
 glib-sys = { git = "https://github.com/gtk-rs/sys" }
 gobject-sys = { git = "https://github.com/gtk-rs/sys" }
 gstreamer-sys = { git = "https://github.com/sdroege/gstreamer-sys" }
